pyproject.toml
[tool.poetry]
name = "reborn-rebalance"
version = "0.7.0"
description = ""
authors = ["Lura Skye <l@veriny.tf>"]
readme = "README.rst"
packages = [{include = "reborn_rebalance", from = "src"}]
license = "CC0"

[tool.poetry.dependencies]
python = "^3.11"
ruamel-yaml = "^0.17.32"
attrs = "^23.1.0"
cattrs = "^23.1.2"
prettyprinter = "^0.18.0"
jinja2 = "^3.1.2"
pillow = "^10.0.0"
tomlkit = "^0.11.8"

[tool.isort]
profile = "black"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
